Desktop/AI_AGENT/Advanced_RAG_CHAT_BOT/requirements.txt
streamlit
langgraph
faiss-cpu
pinecone-client
google-generativeai
langchain
langchain-google-genai
langchain-community
python-dotenv
numpy
pandas
